requirements.txt
# OCR Application Dependencies
# Install with: pip install -r requirements.txt

# Core ML Dependencies
# IMPORTANT: For GPU support, install PyTorch with CUDA before other dependencies:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# Then install the rest:
#   pip install -r requirements.txt
#
# If you only have CPU, the default installation will work fine.

git+https://github.com/huggingface/transformers
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
pillow>=10.0.0
pypdfium2>=4.0.0

# CLI Enhancement
tqdm>=4.65.0

# Utilities
numpy>=1.24.0
huggingface_hub>=0.19.0

# Web Backend (FastAPI Server)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
websockets>=12.0
